+++ acbs/magic.py
