+++ acbs/utils.py
@@ -2,7 +2,7 @@ import subprocess
 import logging
 import sys
 import traceback
-from lib.acbs_const import acbs_const
+from acbs.const import acbs_const
 
 
 class acbs_utils(object):
